src/requirements.txt
Flask~=3.0.3
httpx~=0.27.0
python-telegram-bot~=21.3
PyYAML~=6.0.1